pyproject.toml
[project]
name = "generativeaiartweb"
version = "0.1.0"
description = ""
authors = [
    {name = "Kai Kogure",email = "weizard@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "fire (>=0.7.0,<0.8.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "protobuf (>=6.30.0,<7.0.0)",
    "webdataset (>=0.2.111,<0.3.0)",
    "safetensors (>=0.5.3,<0.6.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "gradio (>=5.20.1,<6.0.0)",
    "flask (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"

